Implemented scalable snapshot sizes

+++ core/src/io/anuke/mindustry/entities/units/UnitDrops.java
@@ -1,14 +1,20 @@
 package io.anuke.mindustry.entities.units;
 
+import io.anuke.mindustry.Vars;
 import io.anuke.mindustry.content.Items;
 import io.anuke.mindustry.gen.CallEntity;
 import io.anuke.mindustry.type.Item;
 import io.anuke.ucore.util.Mathf;
 
 public class UnitDrops {
+    private static final int maxItems = 200;
     private static Item[] dropTable;
 
     public static void dropItems(BaseUnit unit){
+        if(Vars.itemGroup.size() > maxItems){
+            return;
+        }
+
         if(dropTable == null){
             dropTable = new Item[]{Items.tungsten, Items.lead, Items.carbide};
         }
